move backend-specific bits into their backend-appropriate modules
