+++ flocker/node/script.py
@@ -372,55 +372,6 @@ class DeployerType(Names):
     block = NamedConstant()
 
 
-from flocker.node.agents.cinder import cinder_api
-from flocker.node.agents.test.blockdevicefactory import (
-    _openstack_auth_from_config
-)
-from keystoneclient.session import Session
-from cinderclient.client import Client as CinderClient
-from novaclient.client import Client as NovaClient
-
-
-def cinder_from_configuration(**config):
-    """
-    Build a ``CinderBlockDeviceAPI`` using configuration and
-    credentials in ``config``.
-    """
-    region = config.pop('region')
-    auth = _openstack_auth_from_config(**config)
-    session = Session(auth=auth)
-    cinder_client = CinderClient(
-        session=session, region_name=region, version=1
-    )
-    nova_client = NovaClient(
-        session=session, region_name=region, version=2
-    )
-
-    return cinder_api(
-        cinder_client=cinder_client,
-        nova_client=nova_client,
-        cluster_id=config['cluster_id']
-    )
-
-from flocker.node.agents.ebs import EBSBlockDeviceAPI, ec2_client
-
-
-def aws_from_configuration(**config):
-    """
-    Build an ``EBSBlockDeviceAPI`` using configuration and credentials
-    in ``config``.
-    """
-    return EBSBlockDeviceAPI(
-        ec2_client=ec2_client(
-            region=config['region'],
-            zone=config['zone'],
-            access_key_id=config['access_key_id'],
-            secret_access_key=config['secret_access_key'],
-        ),
-        cluster_id=config['cluster_id']
-    )
-
-
 class BackendDescription(PRecord):
     """
     Represent one kind of storage backend we might be able to use.
@@ -449,6 +400,9 @@ class BackendDescription(PRecord):
             value in DeployerType.iterconstants(), "Unknown deployer_type"
         ),
     )
+
+from .agents.cinder import cinder_from_configuration
+from .agents.ebs import aws_from_configuration
 
 # These structures should be created dynamically to handle plug-ins
 _DEFAULT_BACKENDS = [
